Renamed package and module
Renamed package "scraper" to "scraping" as this holds all the scraping
modules.
renamed module "scraping" to "scraper" as this is the high level scraper
module that other specific scrapers inherit from.
Unfortunately this has had import effects in all files but I think it
makes more sense in the long run.